+++ src/main/java/AküListesi.java
@@ -4,11 +4,11 @@ import java.util.ArrayList;
 public class AküListesi{
   private ArrayList<Akü> aküler= new ArrayList<Akü>();
   public AküListesi(Scanner scanner){
-    int voltaj;
+    double voltaj;
     String sağlık;
     while(true){
       System.out.println("Aküler için değerleri giriniz, Voltaj: (Eğer başka akü yoksa 0 giriniz)");
-      voltaj=Integer.valueOf(scanner.nextLine());
+      voltaj=Double.valueOf(scanner.nextLine());
       if (voltaj==0){
         break;
       }
